Add files via upload
The first part of the code finds the optimal low and high bids, this is found by trying to maximize the expectation of profit. The second part accounts for the clause that states that "For your second bid, they also take into account the average of the second bids by other traders in the archipelago. They’ll trade with you when your offer is above the average of all second bids. But if you end up under the average, the probability of a deal decreases rapidly. ". Hence, the scaling factor as shown on the wiki is added in.